+++ news.md
@@ -5,6 +5,8 @@
 - Added the bilateral time-product-dummy index method. This uses the Kennedy (1981) bias adjustment. 
 - The bilateral time-product-dummy method added as a valid bilateral index to use within the GEKS multilateral index number method.
 - Added the Geary-Khamis multilateral method
+- Added the function CESData() that calculates sample datasets using any value of sigma, assuming the data and methodology in Diewert and Fox (2017)
+- Added basePeriod parameter to priceIndex() and quantityIndex() so it is now possible to choose the base period when using output = "fixedbase" for bilateral indices. 
 
 ### Bug fixes