+++ news.md
@@ -3,6 +3,7 @@
 ### Improvements
 
 - Added the bilateral time-product-dummy index method. This uses the Kennedy (1981) bias adjustment. 
+- Added the Geary-Khamis bilateral method.
 - The bilateral time-product-dummy method added as a valid bilateral index to use within the GEKS multilateral index number method.
 - Added the Geary-Khamis multilateral method
 - Added the function CESData() that calculates sample datasets using any value of sigma, assuming the data and methodology in Diewert and Fox (2017)