+++ news.md
@@ -2,8 +2,8 @@
 
 ### Improvements
 
-- Added the bilateral time-product-dummy index method. This optionally uses the Kennedy (1981) bias adjustment. 
-- Added the Geary-Khamis bilateral method.
+- Added the bilateral time-product-dummy index number method. This optionally uses the Kennedy (1981) bias adjustment. 
+- Added the Geary-Khamis bilateral index number method.
 - The bilateral time-product-dummy method added as a valid bilateral index to use within the GEKS multilateral index number method.
 - Added the Geary-Khamis and weighted time-product-dummy multilateral methods.
 - Added the Fixed Base Expanding Window and Fixed Base Moving Window splicing methods.