+++ news.md
@@ -1,4 +1,4 @@
-## Package update 0.1.3 -> 0.2
+## Package update 0.1.3 -> 0.2.0
 
 ### Improvements